+++ build-plugin.js
@@ -10,6 +10,9 @@ var handler = function (compileStep) {
         moduleId = compileStep.packageName + '/' + moduleId;
     }
 
+    // add meteor:// protocol to unify behavior between client/server/packages
+    moduleId = 'meteor://' + moduleId;
+
     var extraWhitelist = ['es6.modules'];
     if(path[1] === 'jsx'){
         extraWhitelist.push('react');