+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 
-## 0.2
+## 0.3
 
-### 0.2.1
+- Update SystemJS to 0.18.4
+- Remove dependency on xhr2
+- Use `meteor://` protocol on modules
+- Alternative import name `foomodule.import` to allow TypeScript support
+- Improve error reporting
+
+## 0.2.1
 
 - Update babel-compiler to 5.6.15
 
-### 0.2.0
+## 0.2.0
 
 - Switch to MDG Babel compiler package